svg/struct/scripted/svg-getIntersectionList-006.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>getIntersectionList() with &lt;use&gt;</title><h:link href="https://svgwg.org/svg2-draft/struct.html#__svg__SVGSVGElement__getIntersectionList" rel="help"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><defs><g id="conn"><circle cx="0" cy="0" r="20"/></g></defs><circle cx="65" cy="50" r="20" fill="#00f"/><use x="50" y="50" fill="red" href="#conn"/><script>function getMatchingLocalNames(x, y) { const svg = document.documentElement; const rect = Object.assign(svg.createSVGRect(), {x, y, width: 1, height: 1}); const elems = svg.getIntersectionList(rect, null); return Array.from(elems).map(e =&gt; e.localName).sort(); } test(t =&gt; { assert_array_equals(getMatchingLocalNames(50, 60), [&apos;circle&apos;, &apos;use&apos;]); assert_array_equals(getMatchingLocalNames(0, 0), []); });</script></svg>
